// src/core/models/xample-dashboard.models.ts
export interface ITest {
  first_name: string;
  last_name: string;
  full_name: string;
  email: string;
  age: number;
  street: string;
  city: string;
  province: string;
  nation: string;
  postal_code: string;
  job_title: string;
  company_name: string;
  company_address: string;
  work_email: string;
}